move _stream_function_call_data into modal._utils.function_utils
